pyproject.toml
[project]
name = "mt5-trading"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "metatrader5>=5.0.5430",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
]
